Update Hyper to 0.10, Multipart to 0.15, Serde to 1.0 and drop rustc-serialize
